Fix for rocket start
The altitude is not reseted at motor arming, contrarly to position. And
if we take_off in a throttle_hold mode like loiter, Alt_hold, Hybrid...
it may happen a rocket start, it means the copter jumps 1-2-3-...m high
like a rocket! And that's because the RelAlt is not reseted to baro_alt.
This diferences make the alt controller send full commands to motors.