Copter: Land_detector reset improvement
For non_auto modes where throttle is controlled by RC_3, this allows to
avoid a fake land_detector reset caused by baro disturbance.
It also allows to reset the land_detection in flight in case of fake
land_detection if we loose the "landed" conditions.
This does not changes the way land_detector is trigered.
TO-DO:
Check if needed to change
}else if ((g.rc_3.control_in != 0 && !failsafe.radio)){
by
}else if (g.rc_3.control_in != 0 && (!failsafe.radio ||
failsafe.rc_override_active)){

The fake_detection_check condition is there to avoid fake
set_land_complete(false); during takeoff
The land_detector = 0; was not called properly so that we needed 50
cumulated iteration and not 50 successive ones to trigger the
land_detection

TO-DO2:
Review the conditions to detect landing (acc/gyro/vel?)