Copter: remove thr check during init_arm_motors
This check is redundant now that we have a check within the arm_check()
function.
Removing this check does raise a slight danger that someone could raise
the throttle after arming but before the gyro and baro calibration has
completed but the delay has been greatly shortened from what it once was
so there is much less danger that someone could approach the vehicle
during the short arming delay.